requirements.txt
langchain
openai
python_dotenv
langchain_openai
streamlit
langchain_community